bsidesnova/requirements.txt
Flask
flask_cors
markdownify
PyMupdf
ollama
playwright
pillow
html2text
requests